backend/requirements.txt
# Fashion Recommendation System Backend Requirements
# Complete dependency list for the entire repository

# Core web framework and API
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
requests==2.31.0

# Deep Learning and Computer Vision
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
timm==0.9.12
opencv-python==4.8.1.78
Pillow==10.0.1

# Machine Learning and Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
imbalanced-learn==0.11.0
scipy>=1.9.0

# Data Visualization (for training/analysis)
matplotlib==3.7.2
seaborn==0.12.2

# Progress bars and utilities
tqdm==4.66.1

# Web Scraping
selenium==4.15.2
webdriver-manager==4.0.1

# LLM Integration - Hugging Face and LangChain
openai>=1.0.0
transformers>=4.21.0
langchain>=0.1.0
langchain-huggingface>=0.0.3
langchain-core>=0.1.0
huggingface-hub>=0.16.0

# Environment and configuration
python-dotenv==1.0.0

# Additional utilities
psutil>=5.9.0

# Note: The following are built-in Python modules (no installation needed):
# json, re, os, subprocess, traceback, gc, sys, io, time, collections, contextlib, base64, uuid